Prometheus: Refactor backend for scalability for other query types
