+++ pkg/tsdb/prometheus/time_series_query_test.go
@@ -14,7 +14,7 @@ import (
 
 var now = time.Now()
 
-func TestPrometheus_formatLeged(t *testing.T) {
+func TestPrometheus_timeSeriesQuery_formatLeged(t *testing.T) {
 	t.Run("converting metric name", func(t *testing.T) {
 		metric := map[p.LabelName]p.LabelValue{
 			p.LabelName("app"):    p.LabelValue("backend"),
@@ -54,7 +54,7 @@ func TestPrometheus_formatLeged(t *testing.T) {
 	})
 }
 
-func TestPrometheus_parseQuery(t *testing.T) {
+func TestPrometheus_timeSeriesQuery_parseTimeSeriesQuery(t *testing.T) {
 	service := Service{
 		intervalCalculator: intervalv2.NewCalculator(),
 	}
@@ -72,7 +72,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, time.Second*30, models[0].Step)
 	})
@@ -91,7 +91,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, time.Second*15, models[0].Step)
 	})
@@ -110,7 +110,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, time.Minute*20, models[0].Step)
 	})
@@ -129,7 +129,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, time.Minute*2, models[0].Step)
 	})
@@ -150,7 +150,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		dsInfo := &DatasourceInfo{
 			TimeInterval: "240s",
 		}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, time.Minute*4, models[0].Step)
 	})
@@ -169,7 +169,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, "rate(ALERTS{job=\"test\" [2m]})", models[0].Expr)
 	})
@@ -188,7 +188,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, "rate(ALERTS{job=\"test\" [120000]})", models[0].Expr)
 	})
@@ -207,7 +207,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, "rate(ALERTS{job=\"test\" [120000]}) + rate(ALERTS{job=\"test\" [2m]})", models[0].Expr)
 	})
@@ -226,7 +226,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, "rate(ALERTS{job=\"test\" [172800s]})", models[0].Expr)
 	})
@@ -245,7 +245,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, "rate(ALERTS{job=\"test\" [172800]})", models[0].Expr)
 	})
@@ -264,7 +264,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, "rate(ALERTS{job=\"test\" [172800000]})", models[0].Expr)
 	})
@@ -283,7 +283,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, "rate(ALERTS{job=\"test\" [1m]})", models[0].Expr)
 	})
@@ -303,7 +303,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, true, models[0].RangeQuery)
 	})
@@ -324,7 +324,7 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, true, models[0].RangeQuery)
 		require.Equal(t, true, models[0].InstantQuery)
@@ -344,13 +344,13 @@ func TestPrometheus_parseQuery(t *testing.T) {
 		}`, timeRange)
 
 		dsInfo := &DatasourceInfo{}
-		models, err := service.parseQuery(query, dsInfo)
+		models, err := service.parseTimeSeriesQuery(query, dsInfo)
 		require.NoError(t, err)
 		require.Equal(t, true, models[0].RangeQuery)
 	})
 }
 
-func TestPrometheus_parseResponse(t *testing.T) {
+func TestPrometheus_parseTimeSeriesResponse(t *testing.T) {
 	t.Run("exemplars response should be sampled and parsed normally", func(t *testing.T) {
 		value := make(map[PrometheusQueryType]interface{})
 		exemplars := []apiv1.ExemplarQueryResult{
@@ -389,7 +389,7 @@ func TestPrometheus_parseResponse(t *testing.T) {
 		query := &PrometheusQuery{
 			LegendFormat: "legend {{app}}",
 		}
-		res, err := parseResponse(value, query)
+		res, err := parseTimeSeriesResponse(value, query)
 		require.NoError(t, err)
 
 		// Test fields
@@ -423,7 +423,7 @@ func TestPrometheus_parseResponse(t *testing.T) {
 		query := &PrometheusQuery{
 			LegendFormat: "legend {{app}}",
 		}
-		res, err := parseResponse(value, query)
+		res, err := parseTimeSeriesResponse(value, query)
 		require.NoError(t, err)
 
 		require.Len(t, res, 1)
@@ -454,7 +454,7 @@ func TestPrometheus_parseResponse(t *testing.T) {
 		query := &PrometheusQuery{
 			LegendFormat: "",
 		}
-		res, err := parseResponse(value, query)
+		res, err := parseTimeSeriesResponse(value, query)
 		require.NoError(t, err)
 
 		var nilPointer *float64
@@ -474,7 +474,7 @@ func TestPrometheus_parseResponse(t *testing.T) {
 		query := &PrometheusQuery{
 			LegendFormat: "legend {{app}}",
 		}
-		res, err := parseResponse(value, query)
+		res, err := parseTimeSeriesResponse(value, query)
 		require.NoError(t, err)
 
 		require.Len(t, res, 1)
@@ -501,7 +501,7 @@ func TestPrometheus_parseResponse(t *testing.T) {
 		}
 
 		query := &PrometheusQuery{}
-		res, err := parseResponse(value, query)
+		res, err := parseTimeSeriesResponse(value, query)
 		require.NoError(t, err)
 
 		require.Len(t, res, 1)
